requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
flower==2.0.1
humanize==4.9.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
pytest==8.0.2
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
